outfit-lab/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
python-telegram-bot==20.3
sqlite3
requests==2.31.0